Move TorrentState to a torrent_state/live
